pyproject.toml
[project]
name = "kaggle-google-agent-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.18.0",
    "marimo>=0.17.7",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "google-adk>=1.18.0",
]
